feat: create postgres db for all
