+++ next-client/src/components/AnimationWrapper.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { motion } from "framer-motion";
+import { useMemo } from "react";
 
 export default function AnimationWrapper({
   children,
